requirements.txt
python-telegram-bot==21.3
httpx>=0.25.2
python-dotenv==1.0.0
langdetect==1.0.9
groq>=0.4.1
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10
